requirements.txt
Django>=4.2,<5.0
graphene-django>=3.0.0
django-filter>=24.0
